Refactor codebase into modular Swift files
